Voicing and Note-Picking (#115)
* [#176696738] XJ has a serviceable voicing algorithm

Based on note-picking test #1 by Mark Stewart
https://docs.google.com/document/d/1kj2v6FaoZN2ztHd44wN4HvSULufleGXw5dd8lYefBig/edit

This test is similar to Test #1 except the following:
  1. The second chord, F, occurs at 1.5 instead of 2.0
  2. The Pad/Stab starting chord is in root position rather than second inversion

Choose correct instrument note based on detail + voicing
https://www.pivotaltracker.com/story/show/176695166

[#176474164] If Sequence has no key/tempo/density inherit from Program

This pad is held for the duration of the C chord as dictated by the main program; it is not re-attacked with each new eighth note

* Main branch

* [#176696738] XJ has a serviceable voicing algorithm

+++ nexus/src/main/java/io/xj/service/nexus/craft/arrangement/ArrangementCraftImpl.java
@@ -151,7 +151,7 @@ public class ArrangementCraftImpl extends FabricationWrapperImpl {
       fabricator.randomlySelectPatternOfSequenceByVoiceAndType(choice, ProgramSequencePattern.Type.Outro);
 
     // compute in and out points, and length # beats for which loop patterns will be required
-    long loopOutPos = (int) maxPos - (outroPattern.map(ProgramSequencePattern::getTotal).orElse(0));
+    double loopOutPos = maxPos - (outroPattern.map(ProgramSequencePattern::getTotal).orElse(0));
 
     // begin at the beginning and fabricate events for the segment of beginning to end
     double curPos = fromPos;
